+++ tests/allTests.c
@@ -13,10 +13,10 @@
 #include "../../sonLib/lib/sonLibCommon.h"
 
 CuSuite* pairwiseAlignmentTestSuite(void);
-CuSuite* multipleAlignerTestSuite(void);
-CuSuite* pairwiseAlignmentLongTestSuite(void);
+//CuSuite* multipleAlignerTestSuite(void);
+//CuSuite* pairwiseAlignmentLongTestSuite(void);
 //CuSuite* shimTestSuite(void);
-//CuSuite* kmerTestSuite(void);
+
 
 int stBaseAlignerRunAllTests(void) {
     CuString *output = CuStringNew();
@@ -25,7 +25,6 @@ int stBaseAlignerRunAllTests(void) {
     //CuSuiteAddSuite(suite, multipleAlignerTestSuite());
     //CuSuiteAddSuite(suite, pairwiseAlignmentLongTestSuite());
     //CuSuiteAddSuite(suite, shimTestSuite());
-    //CuSuiteAddSuite(suite, kmerTestSuite());
     CuSuiteRun(suite);
     CuSuiteSummary(suite, output);
     CuSuiteDetails(suite, output);
